TECH: Refactor connector structure
