+++ src/react/features/connectors/folder/client/mockResponses/GET_ROOM.json
