+++ src/react/features/connectors/folder/client/mockResponses/GET_ROOM2.json
