+++ src/react/utils/cucumber/support/data/expectLinksInDiory.js
@@ -1,5 +1,7 @@
 export const expectLinksInDiory = (idPrefix, id, links) => {
-  const { diograph } = require('../../../../features/connectors/folder/hooks/mockResponses/GET_ROOM.json')
+  const {
+    diograph,
+  } = require('../../../../features/connectors/folder/client/mockResponses/GET_ROOM.json')
   const diory = diograph[idPrefix + id]
   expect(Object.keys(diory.links).length).toBe(links)
 }
